requirements.txt
slackclient
sseclient-py
python-dotenv
requests
flask